moves EmailForm from EditMode dir to GroupManagement (root)
